requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
accelerate>=0.20.0
evaluate>=0.4.0
pyyaml>=6.0
rich>=13.0.0
sentencepiece>=0.1.99
sacremoses>=0.0.53

